+++ src/Resources/contao/languages/de/tl_settings.php
@@ -12,3 +12,5 @@ $GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_host'] = array('Host', 'Host-
 $GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_db'] = array('Datenbank', 'Name der Ergebnisdienst-Datenbank');
 $GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_user'] = array('Benutzer', 'Benutzer der Ergebnisdienst-Datenbank');
 $GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_pass'] = array('Passwort', 'Passwort der Ergebnisdienst-Datenbank');
+$GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_newsletter'] = array('Newsletter-Verteiler', 'Newsletter-Verteiler auswählen, dessen Empfänger automatisch verwaltet werden sollen.');
+$GLOBALS['TL_LANG']['tl_settings']['schiedsrichter_newsletterMail'] = array('Newsletter Zwangs-E-Mail', 'Die hier angegebene Adresse bleibt automatisch Empfänger im ausgewählten Newsletter-Verteiler.');